Translate client_lens_run and find_definitions
